Relocation of authenticator and product_finder
