+++ clothing_compare/clothing_app/external_apis/product_finder.py
@@ -3,7 +3,7 @@ import json
 import requests
 import os
 
-from clothing_app.authenticator import authenticate
+from clothing_compare.clothing_app.external_apis.authenticator import authenticate
 
 URL = "https://api.inditex.com/pubvsearch/products"
 USER = os.environ.get("USER_INDITEX_API")
